Add support for specifying fork owner and name and detect the current user based on the token
The token provided may be different from the repository owner, this fixes an issue where the a fork may be used that the token doesn't have access to.  An example would be if a token from a bot account is used.

Also added are the fork_owner, and fork_name inputs which let the user specify which existing fork they may want to use.  These are optional.